moved site_templates() function
