+++ src/Microweber/Application.php
@@ -31,6 +31,7 @@ namespace Microweber;
  * @property \Microweber\Providers\Template $template
  * @property \Microweber\Providers\Event $event_manager
  * @property \Microweber\Providers\ConfigurationManager $config_manager
+ * @property \Microweber\Providers\TemplateManager $template_manager
  */
 class Application
 {
