+++ src/Microweber/Views/install.php
@@ -369,7 +369,7 @@
                                        onblur="prefix_add(this)"/>
                             </div>
                             <?php
-                            $templates = mw()->template->site_templates();
+                            $templates = mw()->template_manager->site_templates();
 
 
                             ?>
